--- src/main/java/com/ll/tourdemonde/MapController.java
@@ -1,22 +1,0 @@
-//package com.ll.tourdemonde;
-//import org.springframework.stereotype.Controller;
-//import org.springframework.web.bind.annotation.*;
-//
-//
-//@Controller
-//public class MapController {
-//    @GetMapping("/map/save")
-//    @ResponseBody
-//    String showMapSave(){
-//        return "map/save";
-//    }
-//
-////    @RequestMapping(value = "/map", method = RequestMethod.GET, produces = "application/json;charset=UTF-8")
-////    public String getKakaoApiFromAddress(@RequestParam("address") String roadFullAddr) {
-////        String apiKey = "8ea199693cc2ce1fcd18ed34d8b241d7";
-////        StringapiUrl = "https://dapi.kakao.com/v2/local/search/address.json";
-////        String jsonString =null;
-////
-////        return jsonString;
-////    }
-//}
