+++ src/main/java/com/ll/tourdemonde/home/controller/HomeController.java
@@ -4,6 +4,7 @@ package com.ll.tourdemonde.home.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
+
 @Controller
 
 public class HomeController {
@@ -13,4 +14,16 @@ public class HomeController {
         return "redirect:/post/list";
 
     }
+
+    @GetMapping("/map")
+//    @ResponseBody
+    String showMap(){
+        return "map";
+    }
+    @GetMapping("/search")
+//    @ResponseBody
+    String showSearch(){
+        return "search";
+    }
+
 }
