+++ src/main/resources/application-test.yml
@@ -1,3 +1,26 @@
+#spring:
+#  profiles:
+#    active: test
+#    include: secret
+#  h2:
+#    console:
+#      enabled: true
+#      path: /h2-console
+#  datasource:
+#    url: jdbc:h2:mem:TdM_dev_test;MODE=MYSQL # DB : memory mode
+#    username: sa
+#    password:
+#    driver-class-name: org.h2.Driver
+#  jpa:
+#    hibernate:
+#      ddl-auto: create
+#logging:
+#  level:
+#    com.ll.tourdemonde: DEBUG
+#    org.hibernate.SQL: DEBUG
+#    org.hibernate.orm.jdbc.bind: TRACE
+#    org.hibernate.orm.jdbc.extract: TRACE
+#    org.springframework.transaction.interceptor: TRACE
 spring:
   output:
     ansi:
