+++ src/main/resources/application.yml
@@ -1,3 +1,48 @@
+#spring:
+#  profiles:
+#    active: dev
+#    include: secret # personal yml
+#  servlet:
+#    multipart:
+#      max-file-size: 50MB
+#      max-request-size: 50MB
+#  threads:
+#    virtual:
+#      enabled: true
+#  datasource:
+#    url: jdbc:h2:./TdM_dev;MODE=MYSQL # DB : TdM_dev
+#    username: sa
+#    password:
+#    driver-class-name: org.h2.Driver
+#  jpa:
+#    hibernate:
+#      ddl-auto: update
+#    properties:
+#      hibernate:
+#        default_batch_fetch_size: 100
+#        format_sql: true
+#        highlight_sql: true
+#        use_sql_comments: true
+#  thymeleaf:
+#    cache: false
+#  mvc:
+#    hiddenmethod:
+#      filter:
+#        enabled: true
+#logging:
+#  level:
+#    com.ll.tourdemonde: DEBUG
+#    org.hibernate.SQL: DEBUG
+#    org.hibernate.orm.jdbc.bind: TRACE
+#    org.hibernate.orm.jdbc.extract: TRACE
+#    org.springframework.transaction.interceptor: TRACE
+#custom:
+#  site:
+#    name: Tour-de-Monde
+#    baseUrl: "http://localhost:${server.port}"
+#  tempDirPath: c:/temp/TdM_dev/temp
+#  genFile:
+#    dirPath: c:/temp/TdM_dev
 spring:
   profiles:
     active: dev
